Invest track guptapiyush 15

+++ springapp/src/main/java/com/examly/springapp/repository/FeedbackRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.examly.springapp.model.Feedback;
 
 @Repository
-public interface FeedbackRepo extends JpaRepository<Feedback,Integer>{
+public interface FeedbackRepo extends JpaRepository<Feedback,Long>{
 
 }
